+++ lib/src/ecpair.dart
@@ -1,7 +1,10 @@
+import 'dart:convert';
 import 'dart:typed_data';
 import 'dart:math';
 import 'package:bip32_defichain/src/utils/ecurve.dart' as ecc;
 import 'package:bip32_defichain/src/utils/wif.dart' as wif;
+import 'package:defichaindart/defichaindart.dart';
+import 'package:hex/hex.dart';
 import 'models/networks.dart';
 
 class ECPair {
@@ -30,6 +33,12 @@ class ECPair {
 
   Uint8List sign(Uint8List hash) {
     return ecc.sign(hash, privateKey!);
+  }
+
+  String signMessage(String message, [NetworkType? network]) {
+    var hash = magicHash(message, network);
+    var secp = ecc.sign(hash, privateKey!);
+    return base64Encode(ecc.encodeSignature(secp, 0, true));
   }
 
   bool verify(Uint8List hash, Uint8List signature) {