+++ lib/src/models/networks.dart
@@ -33,7 +33,7 @@ final testnet = NetworkType(
     messagePrefix: '\x18Bitcoin Signed Message:\n', bech32: 'tb', bip32: Bip32Type(public: 0x043587cf, private: 0x04358394), pubKeyHash: 0x6f, scriptHash: 0xc4, wif: 0xef);
 
 final defichain = NetworkType(
-    messagePrefix: '\x18Defichain Signed Message:\n', bech32: 'df', bip32: Bip32Type(public: 0x0488B21E, private: 0x0488ADE4), pubKeyHash: 0x12, scriptHash: 0x5a, wif: 0x80);
+    messagePrefix: '\x15Defi Signed Message:\n', bech32: 'df', bip32: Bip32Type(public: 0x0488B21E, private: 0x0488ADE4), pubKeyHash: 0x12, scriptHash: 0x5a, wif: 0x80);
 
-final defichain_testnet = NetworkType(
-    messagePrefix: '\x18Defichain Signed Message:\n', bech32: 'tf', bip32: Bip32Type(public: 0x043587cf, private: 0x04358394), pubKeyHash: 0xf, scriptHash: 0x80, wif: 0xef);
+final defichain_testnet =
+    NetworkType(messagePrefix: '\x15Defi Signed Message:\n', bech32: 'tf', bip32: Bip32Type(public: 0x043587cf, private: 0x04358394), pubKeyHash: 0xf, scriptHash: 0x80, wif: 0xef);